requirements.txt
# requirements.txt

Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
reportlab==4.0.7
Pillow==10.1.0
python-dotenv==1.0.0
flasgger==0.9.7b0
PyJWT==2.8.0
psycopg2-binary==2.9.9
pytest==7.4.4
pytest-flask==1.3.0
